+++ package.json
@@ -13,7 +13,8 @@
     "@vueuse/router": "^9.3.1",
     "airtable": "^0.11.5",
     "diff": "^5.1.0",
-    "vue": "^3.2.37"
+    "vue": "^3.2.37",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^3.1.0",
